Initial version of db-bead/jbead files.
This is the initial version of the formerly db-bead
and now jbead files. The source code is still C++, but
the project structure is ready for transcribing the code
to java.

The images are converted from windows style BMP to PNG
with appropriate transparencies.

The eclipse project/classpath files are checked in as
well as the .preferences folder. This should help with
consistent handling on different machines/platforms.