Partial transliteration of TBeadForm to java
Since TBeadForm is a very large class containing almost all
of the code of the project, this cannot be done in one step.
Thus, this is a partial transliteration and the code is not
even compiling right now.

But some important things are already done. For example,
the drawing code for the draft, normal and simulation views
is now extracted in the classes DraftPanel, NormalPanel and
SimulationPanel.

Also done is the conversion of the registry based settings
to Preferences based settings. On windows this will also use
the registry but on other platforms (linux) a file based
mechanism is used. (This is provided by java and the backend
storage is transparent for the project.)

Finally, the ColorIcon class implements an icon that does
not show an image, but instead is solidly filled with a
color. This is used in the toolbar for showing/editing the
used colors.

There is still lot of work to do and I suppose even when I
reach a compilable state I will have to hunt lots of bugs
because the code behaves differently than before. I expect
this in particular in the drawing code.