Adjusted the coding conventions to standard java style.
I went through the code and changed field and method names to
conform to java conventions. I also refactored some eggregious
stuff, like the opendialog/closedialog abomination. Also, the
language setting stuff uses more appropriate names now.

There is still a lot of work to do, but I like the progress.
Of course, the program just compiles but does not actually
run, because I still miss essential things, for example the
GUI building code (this was in dfm files before, now it will
be in the java code explicitely).