Add ImageFactory and use it to set the imageicon of all windows.
The ImageFactory allows to load images and icons from the resources.

It is used in the constructor of all forms and dilaogs to set
the imageicon to the jbead icon.

I also rename the icons from dbbead to jbead.

The ImageFactory can be used in other places too. This is work
to be done.