Started transition to actions and resource bundles
The menus file, edit and view are already changed, more
or less. The others are to be done.

The toolbar is partially converted.

The explicit language changing code was partially thrown
out, because this is better handled implicitly using the
system locale. Should the need to explicitly change the
language arise, then another solution will be created.

Also some names are now more consistently used. But there
is still a lot of this to be done.