Make file format backwards compatible and handle repaints
It is now possible to save a file with db-bead and open it
with jbead. There is one unfortunate thing: the background
color is some kind of special color and has now a special
read handler.

Now the first color is treated like any other color. It just
happens to be the background color at the start. This is
a much cleaner and, I think, better design.