Add ModelListener and fire events if the model changes.
This is for now only preparation. The next step will be to remove the
repaints and rely on the model listener events. For this the form
needs to become a model listener (or have one).

This will enable a much cleaner separation, because various code
can then move into the model, because no direct dependencies on the
GUI exist anymore.