Refactor bead painting code in corrected panel
The whole redraw uses now getFullRect and iterates over this.
This allows to terminate the redraw exactly when the points
get invisible. Due to the correction stuff, this was a problem
before, resulting in non-redraw of one or two lines at the top.

The single point redraw is now vastly simplified and uses points
exclusively.

The same refactoring can be applied in the simulated panel.