Fix some bugs with bead repaint
I added a getRect method to Model and BeadField, that takes
starty and endy. This allows to start with the current scroll
and break when we reach a point above the top.

I hope this is now good. I still cannot really test it since
the scrolling does not work yet! I expect to have some more
problems with scroll and repaints when I manage to implement
the basic scrolling mechanism.