Make model height adjustable
Add a new dialog for entering the model height. Create actions
for height and width change. Use the actions to build the menu.
Use localized strings in the menu/actions.

Currently, this is only a partial change, since the file format
does not support specification of the height. So if I change the
height and save the file, then I cannot load it again! This needs
to be changed with the v2 file format and there needs to be a
legacy handling of this for the v1 file format (only store default
legacy height in case of v1).