Make setWidth non-destructive
Changing the width/circumference of the pattern now changes each
line/round of beads. This retains the already existing pattern and
just adds some empty columns. Currently, these columns are at the
right of the draft added or removed. Later on it might be possible
to specify where to insert the columns.