Fix corrupt dbb save due to incorrect field length
The default, legacy field length is strange. It is not
a proper multiple of the circumference. In case of the
legacy dbb file format, we have to retain this state.

When we set the width by calling BeadField.setWidth, we
reset the field length and now correctly to the
appropriate length.

This fix does not call setWidth if it is not needed
thus restoring proper saving of the default patterns.

But there is, of course, a further fix needed, because
most users will want to change the width and then the
same problem occurs, if the dbb format is used. Thus
in this case we should always retain the field length
and never change it.