Handle legacy dbb format with non-default pattern width.
Now, there is always the same length of field written to the file.
If the field is shorter, then it is filled up with 0 bytes, if it
is longer then it is truncated. But this would only affect the
very topmost row of the legacy dbb format and this row should not
be visible at all.