Move non-standard field length handling to DbbMemento
The DbbMemento is responsible for storing and loading the legacy
dbb file format. It now writes always DEFAULT_SIZE large fields
and now it reads also always DEFAULT_SIZE fields but makes the
fields in memory slightly larger so that it contains the whole
width * height rectangle. This allows to handle the field uniformly
in the rest of the code. This strange limitation is now (rightfully)
contained in the legacy specific DbbMemento.