Add combined file filter for open and update filters using extensions
A new combined file filter is used when opening files. This filter
shows both legacy dbb and new jbb files, thus easing the handling
of the files.

When opening or saving files, the fileformat is now determined
primarily according to the ending of the chosen file. If the file
does not have an ending or has an unknown one, then the chosen
file filter is consulted.

I hope that this more or less takes care of the dual-format
problems.