Add default values and prepare version upgrades
Missing values are now provided using default values.

If the jbead file format is in an earlier version, then an upgrade
function can be written to convert the data to the new format.
Saving is always done in the latest version (except when choosing
to save in the legacy dbb file format).