Split ReportPrinter into ReportInfosPrinter and BeadListPrinter
Move the ReportInfosPrinter to the first column, leave the
BeadListPrinter as the last columns.

This makes the printing code a lot cleaner and easier. Also the
printout is better for very large bead lists. For short bead lists
it is not so good, because under the report infos is empty space
where the short bead list would fit without problems.

There will be further changes, I suppose, but this is certainly
an improvement.