Fire model change event when the file is set in the Model.
This results in a redraw of the report so that after a save
operation, the filename aka pattern name is updated.

Using the model change event is maybe not ideal, because only the
file changed, not the whole model. But adding a specific file
changed event would be a little bit overkill.