Add popup menu to color palette
This allows to

- select the color
- edit the color
- set the color as the background color

The last option sets the color value in the color with index 0.
This is implicitely the background color. This is a very convenient
way to quickly set the background color with one of the colors
in the palette. Since the background color can also be manually
edited (double-click or popup-menu/edit color) we have it both
ways.