Rework draft printer so that only one col is printed if repeat allows it.
Thus, if the repeat is less than one column high, then a full column
is printed. Provided that the pattern is at least one column high.
If the repeat is higher than one column, then as many columns are
printed as is needed to show the full repeat.

If the pattern is less than one column high, then only a partial
column is printed. That is, there is never automatically filled in
pattern printed.

This scheme captures many use cases, I think, and uses much less
printing space. Now it is easily possible, to design a pattern for
the full planned height. Then the report infos show statistics for
the full height, but only the repeat (filled up to full columns)
is printed.

The same behaviour is planned for corrected and simulation printer.