Improve error handling when loading files
Since the files are first loaded into a memento, the current
model and data stays intact until the load basically succeeded.
Only if the file successfully could be loaded into the memento
is is applied to model and form. Thus, in case of a load error
we just do nothing (apart from displaying an appropriate message)
and the old pattern remains intact.