Add build versioning tools
For each build using ant, the build number is incremented (bumped)
and the version.txt and the launch4j config are patched.

There are two new targets in build.xml, bumpminor and bumpmajor
that are used to increase the minor and major versions, respectively.

After a build, the changed version files and config files have to
be committed manually.