Introduce explicit app object, handle print and file open args
There is now an explicit JBeadApp class that contains the static
main method. This allows easier refactoring of the BeadForm class.

It is now possible to open a file by providing its filepath on
the command line. If the first argument is "/p" or "-p" and then
the second argumnent is the file path, then the file is printed.

These two argument processing capabilities allow easy integration
into the windows shell (double-click to open, right-click + print
for printing).