Persists paper type and orientation page settings
The paper type (A4, Letter, ...)  and the page orientation
are now stored in the print category of the settings. Thus
they are persistent.

Other settings, for example the borders and the tray source
are currently ignored.

Also, some more exotic paper sizes are not supported. Fully
supported are all ISO Ax (A0, ..., A10) and the named ones
like Letter, Legal, Executive. The B-Series, C-Series, the
envelopes and more exotic sizes are not handled and fall back
to the default paper size of the printer.