Make jbead java 5 compatible
This allows to run jbead on older macs or on macs where java 6
is not (yet) installed.

I do not know whether this is a real benefit, but since it was
quite easy to accomplish (apart from one or two a little messier
things) I just did it.

Also, I updated the java.icns for the mac, renamed dbbead.ico
to jbead.ico and updated the Info.plist.