Fix issue with calculation of empty used rect
The report panel needs the used rect to calculate used beads etc.
There was an ArrayOutOfBoundsException in case of an empty pattern.
The reason for this was that the Rect includes its end points. And
so it is a little bit difficult to designate an empty rect. The
usual (0,0)-(0,0) does not work, obviously. So I used (0,0)-(-1,-1).
But RectIterator sorts the coordinates to make the rect canonical.
Thus it let the empty rect be iterated and returned (-1,-1) as the
first point, thus triggering the exception.

The fix is to check for begin coordinates < 0 and then not allowing
to iterate.

A better approach (long term) would be to rewrite Rect to not
include the end point. This is the canonical approach in java.
But I am hesitant, because this would entail rather invasive
changes, I suppose.