Fixed the color repeat calculation
A user reported that the bead list was generated wrongly. Actually
the color repeat was less than the full pattern, because the pattern
started and ended with the same colored beads. Thus, the last beads
were taken to be part of the next repeat. But, since we only check
in one repeat (previously calculated) for color repeat, this did
not change, even if multiple repeats were drawn.

Now, the color repeat is always calculated using the full used
height. This might be a little slower, but it is more accurate.
When you have more than one repetition of the pattern is should
be correct. If only one repetition of the pattern is drawn, then
the same effect occurs. But since this might be required it cannot
be easily prevented. It is just an ambigous case.

I recommend to draw more than one repetition of the pattern.