Add proper comparison of versions
The older, lexicographic comparison fails if the digits of the
buildnumber or of the minor number increase, i.e. 1.0.2 would
be larger than 1.0.10.

The new version splits the version string into parts and does
proper, numeric comparison of each part in turn.

I also took the opportunity to add some unit tests.