Refactor update checker code
The code that checks the version is now in its own class called
VersionChecker. It takes a VersionListener instance and notifies
it about the results. This allows to simplify the code in the
InfoUpdateCheckAction significantly and also to reuse the version
checker code in other places. It is planned to perform a silent
version check whenever jbead is started.