Add unit test for VersionChecker class
Because the VersionChecker class is asynchronous and uses a
callback, the unit test is slightly complicated. It involves
a result string array that is synchronized on and that receives
the result of the check. The main test case does poll on the
array until it's first element is non-null. Then the assert
can happen.