Fix bug B0004 by switching background and selected color
To avoid the potential problem of having a color twice
in the palette, setting a given color as the background
color switches the color with the current background color.

This has the effect, that any beads that used the selected
color are now in the old background color.

This fix does not include a check when manually changing
the color. This would be difficult, because not only exact
matches are errorprone, but slightly different color with
almost the same appearance. But I think that this is a
minor problem because users would probably rather use an
existing color instead of redefining it.