Add basic support for bead symbols, but not yet when printing.
Beads can have a color and a symbol. Both properties can be
enabled or disabled. In the draft and the corrected view both
are drawable. In the simulation, only the color is used.

Missing is support for bead symbols in the print out.

Also, the bead printing code should be centralized and
generalized.