Ensure that the correct file extension is used.
Before this change, it was possible to use arbitrary pattern
file names, e.g. pattern.dat. There were cases where users
saved the pattern first as pattern.jbb, then changed the pattern,
saved it again as pattern.dat. Later on they wanted to load
the pattern again, but since the load dialog by default only
shows .dbb and .jbb files, only the previous version was visible
and the user got the impression, that jbead lost the changes.

This provides the correct extension for the fileformat used to
save the pattern and thus prevents this problem.