Add support for color transparency (alpha channel)
As it turns out, starting with java 7, there can be
set the alpha channel of colors in the color chooser
dialog. Jbead did not save the alpha value. This fixes
this.

Unfortunately, any pattern saved with alpha colors
before this fix, cannot be restored. If it is loaded
the colors are with default alpha (opaque). But if
the colors are changed again and the pattern is saved
again, then the alpha values will be retained.

This also works with java 5 and 6, albeit they cannot
change alpha in the color chooser dialog. But it is
possible to edit such a color to another one with opaque
alpha.