French Translation. Creation of  jbead_fr.properties
Here is a proposal for a French translation of your wonderful jbead ! I hope it corresponds to your convenience. Here I explain my choices.

Regarding the views, the term “patron” is more appropriate than “draft”. Indeed, I prefer to use the same word as in sewing. And "patron" is indeed the translation of "pattern". But the real pattern is the view at the left, while the second view is only a representation of the rope. And the third view is also a representation.

Consequently, i translate "Corrected" par "vue plane" : in French, "corrected" means that errors should be corrected or the view is not correctly diplayed. Instead, "vue plane" (plane view) fits better with the idea that we have a view like a planisphere.


Line 40-42 : Exit means Quit. In French "Quitter". So it would be better to use the Q letter. But as a non-programmer i prefer not replace the X letter. Reversely, CTRL+Y and CTRL+Z are the same combination for a french user. (lines 56 and 61)